⚡️ Improve: ProductAggregationService 로직 추가
상품 집계 시 상품 순위를 업데이트
