+++ src/main/kotlin/kr/hhplus/be/server/product/domain/stats/ProductSalesAggregationDailyRepository.kt
@@ -1,7 +1,11 @@
 package kr.hhplus.be.server.product.domain.stats
 
+import org.springframework.data.repository.query.Param
+import java.time.LocalDate
+
 interface ProductSalesAggregationDailyRepository {
     fun save(entity: ProductSalesAggregationDaily): ProductSalesAggregationDaily
     fun findById(id: ProductSalesAggregationDailyId): ProductSalesAggregationDaily?
     fun findAll(ids: List<ProductSalesAggregationDailyId>): List<ProductSalesAggregationDaily>
+    fun findTopProductsByDay(salesDay: LocalDate, limit: Int): List<Map<String, Any>>
 }
