+++ src/main/kotlin/kr/hhplus/be/server/product/infrastructure/JpaPopularProductsDailyRepository.kt
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.product.infrastructure
+
+import kr.hhplus.be.server.product.domain.stats.PopularProductDailyId
+import kr.hhplus.be.server.product.domain.stats.PopularProductsDaily
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface JpaPopularProductsDailyRepository : JpaRepository<PopularProductsDaily, PopularProductDailyId> {
+
+}
